Add a more apparent way of contacting me
This is for people who want to change their account but don't remember their password.

Note that you will always need to be able to prove the account you want to modify is yours. In case of losing the toxme.io password, this means you must be able to log into this account and add me from it.